backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
httpx>=0.27.0
fastf1>=3.3.0
